+++ main/src/Model/Bridge/Config.php
@@ -51,12 +51,10 @@ class Config implements ConfigInterface
      * @var DirectoryList
      */
     private $directoryList;
-
     /**
      * @var ScopeConfigInterface
      */
     private $scopeConfig;
-
     /**
      * @var GeneralConfig
      */
@@ -211,7 +209,9 @@ class Config implements ConfigInterface
                 $this->_getConfigFlag($prefix . 'show_complete_category_path'),
                 $this->_getConfigFlag($prefix . 'category_link_type'),
                 (array)@unserialize($this->_getConfig($prefix . 'attribute_filter_suggestions')),
-                $this->_getConfigFlag($prefix . 'show_outofstock')
+                $this->_getConfigFlag($prefix . 'show_outofstock'),
+                $this->_getConfigFlag($prefix . 'fuzzy_is_active_for_categories'),
+                (float)$this->_getConfig($prefix . 'fuzzy_sensitivity_for_categories')
             );
         }
         return $this->autosuggest;